+++ org.mixedrealitytoolkit.core/Editor/Editors/BaseInteractableEditor.cs
@@ -3,7 +3,7 @@
 
 using System.Collections.Generic;
 using UnityEditor;
-using UnityEditor.XR.Interaction.Toolkit;
+using UnityEditor.XR.Interaction.Toolkit.Interactables;
 using UnityEngine;
 
 namespace MixedReality.Toolkit.Editor
@@ -12,6 +12,7 @@ namespace MixedReality.Toolkit.Editor
     /// A base class for custom <see cref="MRTKBaseInteractable"/> editors.
     /// </summary>
     [CustomEditor(typeof(MRTKBaseInteractable), true)]
+    [CanEditMultipleObjects]
     public class BaseInteractableEditor : XRBaseInteractableEditor
     {
         private SerializedProperty isGazeHovered;