+++ org.mixedrealitytoolkit.spatialmanipulation/Solvers/TapToPlace.cs
@@ -6,7 +6,9 @@ using Unity.Profiling;
 using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.InputSystem;
+using UnityEngine.InputSystem.XR;
 using UnityEngine.XR.Interaction.Toolkit;
+using UnityEngine.XR.Interaction.Toolkit.Interactors;
 using UnityPhysics = UnityEngine.Physics;
 
 namespace MixedReality.Toolkit.SpatialManipulation
@@ -554,7 +556,7 @@ namespace MixedReality.Toolkit.SpatialManipulation
             interactionManager.GetRegisteredInteractors(interactorsCache);
             foreach (IXRInteractor interactor in interactorsCache)
             {
-                if (interactor is XRBaseControllerInteractor controllerInteractor &&
+                if (interactor is XRBaseInputInteractor controllerInteractor &&
                     controllerInteractor.xrController is ActionBasedController actionController)
                 {
                     actionController.selectAction.action.performed += StopPlacementViaPerformedAction;
@@ -575,7 +577,7 @@ namespace MixedReality.Toolkit.SpatialManipulation
             {
                 foreach (IXRInteractor interactor in interactorsCache)
                 {
-                    if (interactor is XRBaseControllerInteractor controllerInteractor &&
+                    if (interactor is XRBaseInputInteractor controllerInteractor &&
                         controllerInteractor.xrController is ActionBasedController actionController)
                     {
                         actionController.selectAction.action.performed -= StopPlacementViaPerformedAction;