+++ org.mixedrealitytoolkit.input/CHANGELOG.md
@@ -2,13 +2,21 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [3.2.1-development] - 2024-03-20
+## [4.0.0-development] - 2024-05-08
+
+### Changed
+
+* Updated package com.unity.xr.interaction.toolkit to 3.0.1
+
+## [3.2.1] - 2024-4-23
 
 ### Fixed
 
+* Fixed missing [CanEditMultipleObject] attributes as per Bug 573 [PR #698](https://github.com/MixedRealityToolkit/MixedRealityToolkit-Unity/pull/698)
 * Add logic to account for a bound but untracked interaction profile. [PR #704](https://github.com/MixedRealityToolkit/MixedRealityToolkit-Unity/pull/704)
+* Reduced package description to support for UPM package publishing in the Unity Asset Store.
 
-## [3.2.0-development] - 2024-03-20
+## [3.2.0] - 2024-03-20
 
 ### Added