+++ org.mixedrealitytoolkit.uxcomponents.noncanvas/CHANGELOG.md
@@ -2,7 +2,19 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [3.1.2-development] - 2024-03-20
+## [4.0.0-development] - 2024-04-18
+
+### Changed
+
+* Updated package com.unity.xr.interaction.toolkit to 3.0.1
+
+## [3.1.3] - 2024-04-17
+
+### Fixed
+
+* Reduced package description to support for UPM package publishing in the Unity Asset Store.
+
+## [3.1.2] - 2024-03-20
 
 ### Fixed